LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-typed-om-expected.txt


PASS anchor() computes to pixels
FAIL anchor-size() computes to pixels assert_equals: expected 25 but got 0
